Create temporary key pair if needed
The ec2uploadimg script is using a helper instance to prepare
the image that is going to be uploaded. For accessing that
instance a ssh key pair is needed, which currently has to be
provided to the script.

With this commit ec2uploadimg creates a temporary key pair, when
no key pair parameter was provided.
That way the number of mandatory options get's reduced.