Extract EC2Setup class
and move temporary key creation to it.
EC2Setup class will create all necessary resources for an upload to EC2
like a VPC with subnet or a security group with SSH port open.
This makes sure that all dependencies are already set up when calling EC2ImageUploader,
otherwise it was necessary to provide these information in as CLI parameters
or add it to the .ec2config file. Now ec2uploadimg can also make sure that everything is
properly setup.