Fix ec2uploadimg when large helper instance used
This patch resolves an issue where a seconday disk is attached
to larger instance sizes and a seconday disk size is non integer.
For example, 139.7G. An Exception has been added to ignore this
error and print a message about the disk being skipped.

This closes https://github.com/SUSE-Enceladus/ec2imgutils/issues/46