Add a check for an image to become 'available'
This patch uses the AWS waiter to add a check to ec2publishimg to
allow it to wait for an an image to become available before it
publishes the image.

This close https://github.com/SUSE-Enceladus/ec2imgutils/issues/45

Add a check for an image to become 'available'

This patch uses the AWS waiter to add a check to ec2publishimg to
allow it to wait for an an image to become available before it
publishes the image.

This close https://github.com/SUSE-Enceladus/ec2imgutils/issues/45

Fix extra blank line