Cleanup ec2removeimg and add confirm option
This patch removes references to ec2publishimg that were
artifacts leftover from the initial script creation.  In
addition, it also implements the --confirm option.  This
option was originally listed as --no-confirm and was
never implemented.  In order to not create a regression,
the option was changed to --confirm.  This keeps the
original default behavior of not asking for a confirmation.
Finally, some logging was changed to just print an error
message when exiting and not a full Traceback.