Update ec2uploadimg to properly seatch for keys in config files.
This patch removes check_account_keys() from ec2uploadimg,
ec2utils.conf, and the units tests. There already exists a check
in all of the tools that is using a different method to check the
account keys. The check_account_keys() is no longer needed.
In addition, updates were done to change the behavior when
a required config value is missing.  Instead of generating a
Traceback, the progam exits with an appropriate error message.